requirements.txt
streamlit
faiss-cpu
fastembed
pymupdf
google-generativeai
python-dotenv
pydantic
tqdm
